Implementadas categorias

+++ db/migrate/20111021223544_add_stock_to_products.rb
@@ -1,5 +1,5 @@
 class AddStockToProducts < ActiveRecord::Migration
   def change
-    add_column :products, :stock, :integer, default: 0
+    add_column :products, :stock, :decimal, default: 0
   end
 end
